app/requirements.txt
asgiref==3.5.0
Django==3.2.6
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.1.0
psycopg2-binary==2.9.1
pytz==2021.3
sqlparse==0.4.2
